requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
db-dtypes==1.0.4
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-api-core==2.10.2
google-auth==2.13.0
google-cloud-bigquery==3.3.5
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
idna==3.4
numpy==1.23.4
packaging==21.3
pandas==1.5.1
proto-plus==1.22.1
protobuf==4.21.9
psycopg2==2.9.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
